feat: AppService와 TxService 분리
